conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.4.0=pyhd8ed1ab_0
aiohttp=3.8.4=py310h8e9501a_0
aiosignal=1.3.1=pyhd8ed1ab_0
appdirs=1.4.4=pyh9f0ad1d_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.2.0=pyh71513ae_0
blinker=1.5=pyhd8ed1ab_0
brotlipy=0.7.0=py310h8e9501a_1005
bzip2=1.0.8=h3422bc3_4
c-ares=1.18.1=h3422bc3_0
ca-certificates=2022.12.7=h4653dfc_0
cachetools=5.3.0=pyhd8ed1ab_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py310h2399d43_3
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=unix_pyhd8ed1ab_2
colorama=0.4.6=pyhd8ed1ab_0
cryptography=39.0.1=py310hfc83b78_0
frozenlist=1.3.3=py310h8e9501a_0
google-auth=2.16.1=pyh1a96a4e_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
grpcio=1.52.1=py310h0f1eb42_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.0.0=pyha770c72_0
joblib=1.2.0=pyhd8ed1ab_0
libabseil=20230125.0=cxx17_hb7217d7_1
libblas=3.9.0=16_osxarm64_openblas
libcblas=3.9.0=16_osxarm64_openblas
libcxx=14.0.6=h2692d47_0
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=11_3_0_hd922786_27
libgfortran5=11.3.0=hdaf2cc0_27
libgrpc=1.52.1=h1208df7_0
liblapack=3.9.0=16_osxarm64_openblas
libllvm11=11.1.0=hfa12f05_5
libopenblas=0.3.21=openmp_hc731615_3
libprotobuf=3.21.12=hb5ab8b9_0
libsqlite=3.40.0=h76d750c_0
libzlib=1.2.13=h03a7124_4
llvm-openmp=15.0.7=h7cfbb63_0
llvmlite=0.39.1=py310h1e34944_1
markdown=3.4.1=pyhd8ed1ab_0
markupsafe=2.1.2=py310h8e9501a_0
multidict=6.0.4=py310h8e9501a_0
ncurses=6.3=h07bb92c_1
numba=0.56.4=py310h3124f1e_0
numpy=1.23.5=py310h5d7c261_0
oauthlib=3.2.2=pyhd8ed1ab_0
openssl=3.0.8=h03a7124_0
packaging=23.0=pyhd8ed1ab_0
pandas=1.5.3=py310h2b830bf_0
pip=23.0=pyhd8ed1ab_0
pooch=1.6.0=pyhd8ed1ab_0
protobuf=4.21.12=py310h0f1eb42_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pyjwt=2.6.0=pyhd8ed1ab_0
pyopenssl=23.0.0=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.9=h3ba56d0_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=3_cp310
pytorch=1.13.1=py3.10_0
pytz=2022.7.1=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
re2=2023.02.01=hb7217d7_0
readline=8.1.2=h46ed386_0
requests=2.28.2=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scikit-learn=1.2.1=py310ha00a7cd_0
scipy=1.10.0=py310ha0d8a01_2
setuptools=67.3.2=pypi_0
six=1.16.0=pyh6c4a22f_0
tensorboard=2.12.0=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=py310hfc83b78_0
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=he1e0b03_0
tqdm=4.64.1=pyhd8ed1ab_0
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022g=h191b570_0
urllib3=1.26.14=pyhd8ed1ab_0
werkzeug=2.2.3=pyhd8ed1ab_0
wheel=0.38.4=pyhd8ed1ab_0
xz=5.2.6=h57fd34a_0
yarl=1.8.2=py310h8e9501a_0
zipp=3.14.0=pyhd8ed1ab_0
zlib=1.2.13=h03a7124_4
